+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ xxxx-xx-xx
 
     * Drop support for python version earlier than 3.9
     * Add ability to "exclude" path in the check and inventory #583
+    * Add support for the special '-' FILE to print to on screen/to stdout in inventory #584
 
 2024-09-16
     Release 11.0.2